feat:review changes;removed comment;add helper hook;
